Move core logic in internals folder
