+++ src/internals/component-compiler.js
