+++ src/internals/component-proxy.js
